requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pydantic
motor
